+++ src/platform.test.ts
@@ -105,7 +105,7 @@ describe('ShellyPlatform', () => {
       matterbridgeDirectory: 'temp',
       matterbridgePluginDirectory: 'temp',
       systemInformation: { ipv4Address: undefined },
-      matterbridgeVersion: '1.6.0',
+      matterbridgeVersion: '1.6.5',
       addBridgedDevice: jest.fn(),
       removeBridgedDevice: jest.fn(),
       removeAllBridgedDevices: jest.fn(),