+++ CHANGELOG.md
@@ -18,6 +18,7 @@ You can also sponsor Tamer here https://buymeacoffee.com/6sjde6vkzl for his inva
 
 ### Changed
 
+- [package]: Requires matterbridge 1.6.5.
 - [package]: Updated dependencies.
 
 <a href="https://www.buymeacoffee.com/luligugithub">